+++ CHANGELOG.md
@@ -1,31 +1,83 @@
 ## 0.0.1
 
-* initial release.
+- initial release.
 
 ## 0.0.2
 
-* MINOR BUG FIXES
+- MINOR BUG FIXES
 
 ## 0.0.3
 
-* MINOR BUG FIXES
+- MINOR BUG FIXES
 
 ## 0.0.4
 
-* MINOR BUG FIXES
+- MINOR BUG FIXES
 
 ## 0.0.5
 
-* MINOR BUG FIXES
+- MINOR BUG FIXES
 
 ## 0.0.6
 
-* Android location fixes
+- Android location fixes
 
 ## 0.0.7
 
-* Added children to map and hide default search bar
+- Added children to map and hide default search bar
 
 ## 0.0.8
 
-* Added routing methods
+- Added routing methods
+
+## 0.0.9
+
+- lower dart dependency
+
+## 0.1.0
+
+- map api tokenized
+
+## 0.1.1
+
+- on galli marker tap fixed
+
+## 0.1.2
+
+- autocomplete distance fix
+
+## 0.1.3
+
+- minor change
+
+## 0.1.4
+
+- fixed bug when showCurrentLocation is false
+
+## 0.1.5
+
+- minor bug fix for navigation
+
+## 0.1.6
+
+- minor bug fix for navigation
+
+## 0.1.7
+
+- readme update
+
+## 0.1.8
+
+- minor fix
+
+## 0.1.9
+
+- dependency fixes
+
+## 0.2.0
+
+- minor fix
+
+## 0.2.1
+
+- map upgrade